Update for latest fastapi release.
